Add GLAD library
To load newer versions of OpenGL, an additional library is
required. GLAD plays that role.

GLAD source and header files were added directly to the source tree.
A custom written CMakeLists.txt file was added as well, which defines
GLAD as a library, defines the necessary includes and links it with
its dependencies