Change include directories definition's position
By replacing `include_directories` in the root `CMakeLists.txt` file with a target-specific `target_include_directories` in the `src/CMakeLists.txt` files, it allows us to be more flexible with future directory structure.

For example, if a mini-library is created for the purposes of the project, its include directory may be a subdirectory of the entire projects include directory.