Move and rename `id` to `intValue`
The function is now defined with other VertexLayout utility functions,
in the `rg::util` namespace.

In the process, discovered and fixed error pertaining to `constexpr`
funcitons. Namely, they do not link, since they are implicitly `inline`.
As a result, the `constexpr` specifier was removed from the
`rg::util::size` function.