Initial camera support
Created a struct `View`, which represents an 'eye' or 'camera' in
graphics programming. The view has a perspective projection.

This struct is fairly minimal; its only real purpose is to generate
the __view__ and __projection__ matrices, which is enough from a
graphics programming perspective, since those are the only
components needed (from a camera) to render a model in the specified
location.

This struct is intended to be used as a part of a camera system,
which would handle the camera's motion and rotation, as well as
the domain for the `View`s values.