Update View misconceptions
The critical issue is that the View class depends on the `glm` library,
and it takes a __vertical field of view__ and an __aspect ratio__ for
its projection calculations, whereas the `View` class sent
`horizontal_fov` as the vertical FOV, and `vertical_fov` as the aspect
ratio.

Coincidentally, radian values of common angles for a field of view do
not look bad as aspect ratios. :)