Update example
The backpack example uses the new `Camera` system, with 4 cameras.
It uses framebuffers to draw on the screen in one of two modes:
1. Single camera view
2. Four-way camera view

The single camera view shows the scene as a single camera sees it.
Switching between cameras on the fly is supported.

The four-way camera view splits the window in quarters, and shows
the scene from each camera in its own quarter.