Optimize rendering
The problem with the previous rendering scheme was that many uniforms
needed to be set every time an object was going to be drawn to a
surface. This resulted in a very noticeable loss in framerate when
objects were drawn on multiple surfaces (4 times the work!).

By using the knowledge that shaders store set uniform values, even
after they've been unbound, most of the code used to set shader
uniforms has been relocated to be set once per frame or once
per frame per surface rendered.