Fix camera direction bug
Yaw and pitch calculation from a direction vector used inverse
trigonometric functions to calculate the yaw and pitch.

This direction vector was normalized, and so the pitch was correct.

The problem is that the yaw calculation is done in 2 dimensions,
and the input vector's projection to the xz plane is not normalized.

The calculation was fixed by creating a new "planar" vector from
the "spatial" one as the projection of the vector to the xz plane,
and normalizing it, and only then performing yaw calculations.